Update AtomsBaseTesting.jl
Added option `:diagonal` to AtomsBaseTesting `test_approx_equal`. The current options are not compatible with Molly.jl (triclinic & DirichletZero) and I would like to test against a valid system.

Also is `DirhichletZero` supposed to be a hard wall or infinite domain or something else?  Doesn't really make a difference for Molly but its not really clear from docs and not something I've ever seen in a molecular simulation context.